LayoutTests/fast/regex/cross-frame-callable-expected.txt

PASS re('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE
DONE
